+++ test/fork/WildcardPermissionKillChain.t.sol
@@ -270,8 +270,7 @@ contract WildcardPermissionKillChain is RevnetForkBase {
             minTokensPaidOut: 0,
             beneficiary: payable(address(REV_DEPLOYER)),
             feeBeneficiary: payable(address(REV_DEPLOYER)), // fee beneficiary for the payout
-            memo: "drain surplus",
-            referralProjectId: 0
+            memo: "drain surplus"
         });
     }
 
@@ -337,8 +336,7 @@ contract WildcardPermissionKillChain is RevnetForkBase {
             minTokensPaidOut: 0,
             beneficiary: payable(address(CT_DEPLOYER)),
             feeBeneficiary: payable(address(CT_DEPLOYER)), // fee beneficiary for the payout
-            memo: "steal funds",
-            referralProjectId: 0
+            memo: "steal funds"
         });
     }